Move ontology-related internal modules into src/scripts
